# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: USDTLuck API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/wallet/transactions:
    get:
      operationId: listCurrentUserWalletLedger
      tags: [users]
      summary: Current user wallet ledger rows (prizes, withdrawals, deposits, bonuses)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Wrapped list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWalletTransactionsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/{userId}/transactions:
    get:
      operationId: getUserTransactions
      tags: [users]
      summary: Get user transaction history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions (admin)
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a deposit or withdrawal transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient balance or invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    UserWalletLedgerRow:
      type: object
      properties:
        id:
          type: integer
        transaction_type:
          type: string
        category:
          type: string
        amount:
          type: number
        reference_type:
          type: string
          nullable: true
        reference_id:
          type: integer
          nullable: true
        description:
          type: string
        balance_after:
          type: number
        created_at:
          type: string
          format: date-time
      required:
        - id
        - transaction_type
        - category
        - amount
        - description
        - balance_after
        - created_at
    UserWalletTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/UserWalletLedgerRow"
      required: [transactions]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        txType:
          type: string
          enum: [deposit, withdraw, reward, pool_entry]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        note:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, txType, amount, status, createdAt]
    CreateTransactionBody:
      type: object
      properties:
        txType:
          type: string
          enum: [deposit, withdraw]
        amount:
          type: number
        note:
          type: string
      required: [txType, amount]
...